fix bug confirming model

+++ train/train_confirming.py
@@ -89,7 +89,7 @@ def train(model, device, train_loader, criterion, optimizer, scheduler, epoch, i
             
             label_pred = decoder(output)
 
-            train_precision = precision_score(np.array(label_pred), np.array(labels.tolist()), average='micro')
+            train_precision = precision_score(np.array(label_pred), np.array(labels.float().tolist()), average='micro')
 
             optimizer.step()
             scheduler.step()
@@ -121,7 +121,7 @@ def test(model, device, test_loader, criterion, iter_meter, experiment, filename
 
                 label_pred = decoder(output)
 
-                test_precision = precision_score(np.array(label_pred), np.array(labels.tolist()), average='micro')
+                test_precision = precision_score(np.array(label_pred), np.array(labels.float().tolist()), average='micro')
 
     experiment.log_metric('test_loss', test_loss, step=iter_meter.get())
     experiment.log_metric('test_precision', test_precision, step=iter_meter.get())
